Complete rewrite: simpler implementation.

+++ src/main/java/math/nyx/core/Fractal.java
@@ -5,17 +5,17 @@ import java.io.Serializable;
 public abstract class Fractal implements Serializable {
 	private static final long serialVersionUID = -3691227743645881425L;
 
-	private long signalDimension;
+	private int signalDimension;
 
 	public abstract Signal decode();
 
 	public abstract Signal decode(float scale);
 
-	public long getSignalDimension() {
+	public int getSignalDimension() {
 		return signalDimension;
 	}
 
-	public void setSignalDimension(long signalDimension) {
+	public void setSignalDimension(int signalDimension) {
 		this.signalDimension = signalDimension;
 	}
 }
